pyproject.toml
[tool.poetry]
name = "dynamodb_isolation_poc"
version = "0.1.0"
description = ""
authors = ["royb <roybenyosef@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"aws_cdk.core" = "^1.22.0"
boto3 = "^1.11.9"
"aws_cdk.aws_dynamodb" = "^1.22.0"
pytest = "^5.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
